+++ test/integration/datapath/datapath-win_test.go
@@ -1,0 +1,179 @@
+package connection
+
+import (
+	"context"
+	"flag"
+	"fmt"
+	"testing"
+
+	"github.com/Azure/azure-container-networking/test/internal/datapath"
+	"github.com/Azure/azure-container-networking/test/internal/k8sutils"
+	"github.com/stretchr/testify/require"
+
+	apiv1 "k8s.io/api/core/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+)
+
+const (
+	WindowsPodYamlPath = "../manifests/datapath/windowspod.yaml"
+	podLabelKey        = "app"
+	podCount           = 2
+)
+
+var (
+	podPrefix         = flag.String("podName", "datapod", "Prefix for test pods")
+	podNamespace      = flag.String("namespace", "datapath-win", "Namespace for test pods")
+	nodeLabelSelector = flag.String("nodeLabelSelector", "agentpool=npwin", "Label selector for nodepool, keep under 6 characters")
+)
+
+/*
+This test assumes that you have the current credentials loaded in your kubeconfig for a
+k8s cluster with a windows nodepool consisting of at least 2 windows nodes.
+
+To run the test use the following command as an example:
+go test -timeout 5m -count=1 test/integration/datapath/datapath-win_test.go -podName=acnpod -nodeLabelSelector="agentpool=npwina" -v
+
+Timeout context is controled by the -timeout flag.
+***Using the default a normal test takes
+
+*/
+
+func TestDatapathWin(t *testing.T) {
+	// var err error
+
+	ctx := context.Background()
+
+	t.Log("Create Clientset")
+	clientset, err := k8sutils.MustGetClientset()
+	if err != nil {
+		require.NoError(t, err, "could not get k8s clientset: %v", err)
+	}
+	t.Log("Get REST config")
+	restConfig := k8sutils.MustGetRestConfig(t)
+
+	// Pod Namespace
+	t.Log("Create Namespace")
+	err = k8sutils.CreateNamespace(ctx, clientset, *podNamespace)
+	createPodFlag := apierrors.IsAlreadyExists(err)
+
+	// Pod Interface Getter
+	t.Log("Pod Interface")
+	podI := clientset.CoreV1().Pods(*podNamespace)
+
+	// Get NodeList WindowsNodePoolName
+	t.Log("Get Nodes")
+	nodes, err := k8sutils.GetNodeListByLabelSelector(ctx, clientset, *nodeLabelSelector)
+	if err != nil {
+		require.NoError(t, err, "could not get k8s node list: %v", err)
+	}
+
+	podLabelSelector := fmt.Sprintf("%s=%s", podLabelKey, *podPrefix)
+	// Checks namespace already exists from previous test
+	if createPodFlag {
+		t.Log("Pod namespace already exists")
+		t.Log("Checking Windows test environment ")
+		for _, node := range nodes.Items {
+
+			pods, err := k8sutils.GetPodsByNode(ctx, clientset, *podNamespace, podLabelSelector, node.Name)
+			if err != nil {
+				require.NoError(t, err, "could not get k8s clientset: %v", err)
+			}
+			if len(pods.Items) < 2 {
+				require.NoError(t, err, "Only %d pods on node %s, requires at least 2 pods", len(pods.Items), node.Name)
+			}
+		}
+		t.Log("Windows test environment ready")
+	} else {
+		t.Log("Creating Windows pods")
+
+		podCounter := 0
+		for _, node := range nodes.Items {
+			for i := 0; i < podCount; i++ {
+				pod, err := k8sutils.MustParsePod(WindowsPodYamlPath)
+				require.NoError(t, err, "Parsing windows pod deployment failed")
+				pod.Spec.NodeSelector = make(map[string]string)
+				pod.Spec.NodeSelector["kubernetes.io/hostname"] = node.ObjectMeta.Name
+				pod.Name = fmt.Sprintf("%s-%d", *podPrefix, podCounter)
+				pod.Namespace = *podNamespace
+				pod.Labels = make(map[string]string)
+				pod.Labels[podLabelKey] = *podPrefix
+				err = k8sutils.MustCreateOrUpdatePod(ctx, podI, pod)
+				require.NoError(t, err, "Creating windows pods failed")
+				err = k8sutils.WaitForPodsRunning(ctx, clientset, pod.Namespace, podLabelSelector)
+				require.NoError(t, err, fmt.Sprintf("Deploying windows Pod: %s failed with error: %v", pod.Name, err))
+				t.Logf("Successfully deployed windows pod: %s", pod.Name)
+				podCounter += 1
+			}
+		}
+		t.Log("Successfully created customer windows pods")
+	}
+
+	t.Run("Windows ping tests pod -> node", func(t *testing.T) {
+		// Windows ping tests between pods and node
+		t.Log("Windows Pod to Host Ping tests")
+		for _, node := range nodes.Items {
+			t.Log("Windows ping tests (2)")
+			// func windowsPodToNode(ctx context.Context, nodeName string, nodeIP string, labelSelector string, rc *restclient.Config) error {
+			nodeIP := ""
+			for _, address := range node.Status.Addresses {
+				if address.Type == "InternalIP" {
+					nodeIP = address.Address
+					// Multiple addresses exist, break once Internal IP found.
+					// Cannot call directly
+					break
+				}
+			}
+
+			err := datapath.WindowsPodToNode(ctx, clientset, node.Name, nodeIP, *podNamespace, podLabelSelector, restConfig)
+			require.NoError(t, err, "Windows pod to node, ping test failed with %+v", err)
+			t.Logf("Windows pod to node, passed for node: %s", node.Name)
+		}
+	})
+
+	t.Run("Windows ping tests pod -> pod", func(t *testing.T) {
+		// Get NodeList WindowsNodePoolName
+		nodes, err := k8sutils.GetNodeListByLabelSelector(ctx, clientset, *nodeLabelSelector)
+		if err != nil {
+			require.NoError(t, err, "could not get k8s node list: %v", err)
+		}
+
+		// Windows pod ping tests
+		for _, node := range nodes.Items {
+			if node.Status.NodeInfo.OperatingSystem == string(apiv1.Windows) {
+				// Pod to pod same node
+				// windowsPodToPodPingTestSameNode(ctx context.Context, clientset *kubernetes.Clientset, nodeName string, labelSelector string, rc *restclient.Config)
+				t.Log("Windows ping tests (1) - Same Node")
+				err := datapath.WindowsPodToPodPingTestSameNode(ctx, clientset, node.Name, *podNamespace, podLabelSelector, restConfig)
+				require.NoError(t, err, "Windows pod to pod, same node, ping test failed with %+v", err)
+				t.Logf("Windows pod to windows pod, same node, passed for node: %s", node.ObjectMeta.Name)
+			}
+		}
+
+		// Pod to pod different node
+		for i := 0; i < len(nodes.Items); i++ {
+			t.Log("Windows ping tests (1) - Different Node")
+			firstNode := nodes.Items[i%2].Name
+			secondNode := nodes.Items[(i+1)%2].Name
+			err = datapath.WindowsPodToPodPingTestDiffNode(ctx, clientset, firstNode, secondNode, *podNamespace, podLabelSelector, restConfig)
+			require.NoError(t, err, "Windows pod to pod, different node, ping test failed with %+v", err)
+			t.Logf("Windows pod to windows pod, different node, passed for node: %s -> %s", firstNode, secondNode)
+
+		}
+	})
+
+	t.Run("Windows url tests pod -> internet", func(t *testing.T) {
+		// From windows pod to curl an URL
+		t.Log("Windows Pod to Internet tests")
+		nodes, err := k8sutils.GetNodeListByLabelSelector(ctx, clientset, *nodeLabelSelector)
+		if err != nil {
+			require.NoError(t, err, "could not get k8s node list: %v", err)
+		}
+		for _, node := range nodes.Items {
+			if node.Status.NodeInfo.OperatingSystem == string(apiv1.Windows) {
+				err := datapath.WindowsPodToInternet(ctx, clientset, node.Name, *podNamespace, podLabelSelector, restConfig)
+				require.NoError(t, err, "Windows pod to internet url %+v", err)
+				t.Logf("Windows pod to Internet url tests")
+			}
+		}
+	})
+}
